+++ src/main/kotlin/com/greenkode/trader/portfolio/Positions.kt
@@ -1,11 +1,12 @@
 package com.greenkode.trader.portfolio
 
 import com.greenkode.trader.domain.Symbol
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
-class Positions(val timestamp: LocalDateTime?, val positions: MutableMap<Symbol, Double>) {
+class Positions(val timestamp: LocalDateTime?, val positions: MutableMap<Symbol, BigDecimal>) {
 
-    fun updateQuantity(symbol: Symbol, quantity: Double) {
+    fun updateQuantity(symbol: Symbol, quantity: BigDecimal) {
         positions[symbol] = quantity
     }
 }
